+++ packages/core-react/src/context/reefKnot.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, ReactNode, useMemo } from 'react';
+import React, { createContext, ReactNode, useMemo, useState } from 'react';
 import { ReefKnotModalContextProvider } from './reefKnotModalContext';
 import type { WalletAdapterData } from '@reef-knot/types';
 
@@ -11,7 +11,10 @@ export interface ReefKnotContextProps {
   children?: ReactNode;
 }
 
-export type ReefKnotContextValue = ReefKnotProviderConfig;
+export type ReefKnotContextValue = ReefKnotProviderConfig & {
+  loadingWalletId: string | null;
+  setLoadingWalletId: React.Dispatch<React.SetStateAction<string | null>>;
+};
 
 export const ReefKnotContext = createContext({} as ReefKnotContextValue);
 
@@ -19,8 +22,19 @@ export const ReefKnotProvider = ({
   config,
   children,
 }: ReefKnotContextProps) => {
+  const [loadingWalletId, setLoadingWalletId] = useState<string | null>(null);
+
+  const contextValue = useMemo(
+    () => ({
+      ...config,
+      loadingWalletId,
+      setLoadingWalletId,
+    }),
+    [config, loadingWalletId],
+  );
+
   return (
-    <ReefKnotContext.Provider value={config}>
+    <ReefKnotContext.Provider value={contextValue}>
       <ReefKnotModalContextProvider>{children}</ReefKnotModalContextProvider>
     </ReefKnotContext.Provider>
   );